requirements.txt
# Core Framework
streamlit>=1.35.0
langchain>=0.2.0
langchain-community>=0.2.0
langchain-groq>=0.1.4

# Embeddings
sentence-transformers>=2.6.0
faiss-cpu>=1.9.0

# YouTube Transcript
youtube-transcript-api>=0.6.2
pytube>=15.0.0

# LLM
groq>=0.4.2

# Utilities
python-dotenv>=1.0.1
pydantic>=2.6.0
tiktoken>=0.6.0